app.py
from flask import Flask
from flask_restx import Api
from dotenv import load_dotenv
import os

from extensions import db, migrate
from controllers.student_controller import student_ns

load_dotenv()

def create_app():
    app = Flask(__name__)

    app.config["SQLALCHEMY_DATABASE_URI"] = (
        f"mysql+pymysql://{os.getenv('DB_USER')}:{os.getenv('DB_PASSWORD')}"
        f"@{os.getenv('DB_HOST')}/{os.getenv('DB_NAME')}"
    )
    app.config["SQLALCHEMY_TRACK_MODIFICATIONS"] = False

    db.init_app(app)
    migrate.init_app(app, db)

    api = Api(app, version="1.0", title="Student API", description="A simple Student Management API")

    api.add_namespace(student_ns)

    return app

app = create_app()

if __name__ == "__main__":
    app.run(debug=True, host="0.0.0.0", port=5000)
